Enhancement #2,"Implement real time clock"
Moved clock constants to a more sensible location. Began adding non volatile storage functions for the clock config, which for now simply includes the last saved time. In the future, this will be used for additional things like personalized colors and effects.

Next step is to finish these NVS functions, and do some testing of the RTC read/write.